Monty Hall Challenge Application.
This program allows you to simulate the Monty Hall Challenge as if you were a contestant.  It allows you the choice to choose between different doors and change the number of doors.  It generates a random prize door based on the number of doors you chose.  It also calculates the probability that you are right in choosing the right door.